Abstracting out MMC functionality
